--- web/.env
@@ -1,5 +1,0 @@
-NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=
-CLERK_SECRET_KEY=
-CLERK_WEBHOOK_SECRET=
-DATABASE_URL=
-WEB_APP_ORIGIN=
