+++ web/src/app/(with-main-layout)/profile/page.tsx
@@ -1,8 +1,10 @@
 "use client";
 import { useUser } from "@clerk/nextjs";
-import { Box } from "@mui/material";
+import { Box, Stack } from "@mui/material";
 import UserProfile from "@/components/user_profile_page/user_profile";
-
+import UserProfileMenu from "@/components/user_profile_page/user-profile-menu";
+import * as React from "react";
+import ReservationsPage from "@/components/user_profile_page/user-reservations";
 export default function Home() {
   const { user, isLoaded } = useUser();
 
@@ -11,6 +13,9 @@ export default function Home() {
   }
   console.log(user);
 
+  const [currentPage, setCurrentPage] = React.useState<'profile' | 'reservations'>('profile');
+  const showProfile = () => setCurrentPage('profile');
+  const showReservations = () => setCurrentPage('reservations');
   return (
     <Box
       sx={{
@@ -19,9 +24,15 @@ export default function Home() {
         margin: 8,
       }}
     >
-
-      <UserProfile />
-
+      <Stack direction={'row'}>
+        <UserProfileMenu showProfile={showProfile} showReservations={showReservations} />
+        {currentPage === "profile" && (
+          <UserProfile />
+        )}
+        {currentPage === "reservations" && (
+          <ReservationsPage />
+        )}        
+      </Stack>
     </Box>
   );
 }
